+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   Prepare application to group projects together for advisory board.
 - new transfer projects now collect the group reference number, used in the
   Prepare application to group projects together for advisory board.
+- the group reference number for a conversion or transfer can now be changed
+  from the project information view.
 
 ## [Release-81][release-81]